+++ beets/dbcore/types.py
@@ -14,6 +14,8 @@
 
 """Representation of type information for DBCore model fields.
 """
+import re
+
 from . import query
 from beets.util import str2bool
 
@@ -164,3 +166,31 @@ class Boolean(Type):
 
     def parse(self, string):
         return str2bool(string)
+
+
+class MusicalKey(String):
+    """String representing the musical key of a song.
+
+    The standard format is C, Cm, C#, C#m, etc.
+    """
+
+    ENHARMONIC = {
+        r'db': 'c#',
+        r'eb': 'd#',
+        r'gb': 'f#',
+        r'ab': 'g#',
+        r'bb': 'a#',
+    }
+
+    def parse(self, key):
+        key = key.lower()
+        for flat, sharp in self.ENHARMONIC:
+            re.sub(flat, sharp, key)
+        re.sub(r'[\W\s]+minor', 'm', key)
+        return key.capitalize()
+
+    def normalize(self, key):
+        if key is None:
+            return None
+        else:
+            return self.parse(key)